refactor: split dashboard-templates.ts into focused modules
Extract the 1,729-line monolith into 5 focused modules:
- dashboard-formatters.ts (45 lines): escapeHtml, DashboardStats, buildDashboardStats
- dashboard-components.ts (71 lines): SVG icons, truncateTitle, renderHealthItems, titleMeta
- dashboard-styles.ts (766 lines): CSS theme variables and component styles
- dashboard-scripts.ts (299 lines): client-side JS for theme toggle, filtering, Chart.js charts
- dashboard-templates.ts (606 lines): main HTML template assembly with re-exports

All existing imports from dashboard-templates.ts continue to work via re-exports.
HTML output is identical. All 1,540 tests pass.

Closes #454
